requirements.txt
streamlit
tensorflow==2.17.1
numpy
Pillow

